API Reader divided in several modules
