+++ apps/explorer/lib/explorer/chain/arbitrum/reader/api/settlement.ex
@@ -1,17 +1,18 @@
-defmodule Explorer.Chain.Arbitrum.Reader.API do
+defmodule Explorer.Chain.Arbitrum.Reader.API.Settlement do
   @moduledoc """
-    Provides API-specific functions for querying Arbitrum-related data from the database.
+    Provides API-specific functions for querying Arbitrum settlement data from the database.
 
     This module contains functions specifically designed for Blockscout's API endpoints
-    that handle Arbitrum-specific functionality. All functions in this module enforce
+    that handle Arbitrum settlement functionality. All functions in this module enforce
     the use of replica databases for read operations by automatically passing the
     `api?: true` option to database queries.
 
     The module includes functions for retrieving:
-    - Cross-chain messages (L1<->L2)
     - L1 batches and their associated transactions
-    - Data Availability (DA) records
+    - Data Availability (DA) records and blobs
     - Batch-related block information
+    - Block confirmations on the parent chain
+    - AnyTrust keysets
 
     Note: If any function from this module needs to be used outside of API handlers,
     it should be moved to `Explorer.Chain.Arbitrum.Reader.Common` with configurable
@@ -26,173 +27,15 @@ defmodule Explorer.Chain.Arbitrum.Reader.API do
     BatchToDaBlob,
     BatchTransaction,
     DaMultiPurposeRecord,
-    L1Batch,
-    Message
+    L1Batch
   }
 
   alias Explorer.Chain.Arbitrum.Reader.Common
   alias Explorer.Chain.Block, as: FullBlock
-  alias Explorer.Chain.{Hash, Log}
-
-  alias Explorer.{Chain, PagingOptions}
   alias Explorer.Chain.Cache.BackgroundMigrations, as: MigrationStatuses
+  alias Explorer.{Chain, PagingOptions}
 
   @api_true [api?: true]
-
-  @doc """
-    Retrieves L2-to-L1 messages initiated by specified transaction.
-
-    The messages are filtered by the originating transaction hash (with any status).
-    In the common case a transaction can initiate several messages.
-
-    ## Parameters
-    - `transaction_hash`: The transaction hash which initiated the messages.
-
-    ## Returns
-    - Instances of `Explorer.Chain.Arbitrum.Message` initiated by the transaction
-      with the given hash, or `[]` if no messages with the given status are found.
-  """
-  @spec l2_to_l1_messages_by_transaction_hash(Hash.Full.t()) :: [Message.t()]
-  def l2_to_l1_messages_by_transaction_hash(transaction_hash) do
-    query =
-      from(msg in Message,
-        where: msg.direction == :from_l2 and msg.originating_transaction_hash == ^transaction_hash,
-        order_by: [desc: msg.message_id]
-      )
-
-    query
-    |> select_repo(@api_true).all()
-  end
-
-  @doc """
-    Retrieves L2-to-L1 message by message id.
-
-    ## Parameters
-    - `message_id`: message ID
-
-    ## Returns
-    - Instance of `Explorer.Chain.Arbitrum.Message` with the provided message id,
-      or nil if message with the given id doesn't exist.
-  """
-  @spec l2_to_l1_message_by_id(non_neg_integer()) :: Message.t() | nil
-  def l2_to_l1_message_by_id(message_id) do
-    query =
-      from(message in Message,
-        where: message.direction == :from_l2 and message.message_id == ^message_id
-      )
-
-    select_repo(@api_true).one(query)
-  end
-
-  @doc """
-    Retrieves the count of cross-chain messages either sent to or from the rollup.
-
-    ## Parameters
-    - `direction`: A string that specifies the message direction; can be "from-rollup" or "to-rollup".
-
-    ## Returns
-    - The total count of cross-chain messages.
-  """
-  @spec messages_count(binary()) :: non_neg_integer()
-  def messages_count(direction) when direction == "from-rollup" do
-    do_messages_count(:from_l2)
-  end
-
-  def messages_count(direction) when direction == "to-rollup" do
-    do_messages_count(:to_l2)
-  end
-
-  # Counts the number of cross-chain messages based on the specified direction.
-  @spec do_messages_count(:from_l2 | :to_l2) :: non_neg_integer()
-  defp do_messages_count(direction) do
-    Message
-    |> where([msg], msg.direction == ^direction)
-    |> select_repo(@api_true).aggregate(:count)
-  end
-
-  @doc """
-    Retrieves cross-chain messages based on the specified direction.
-
-    This function constructs and executes a query to retrieve messages either sent
-    to or from the rollup layer, applying pagination options. These options dictate
-    not only the number of items to retrieve but also how many items to skip from
-    the top.
-
-    ## Parameters
-    - `direction`: A string that can be "from-rollup" or "to-rollup", translated internally to `:from_l2` or `:to_l2`.
-    - `options`: A keyword list which may contain `paging_options` specifying pagination details
-
-    ## Returns
-    - A list of `Explorer.Chain.Arbitrum.Message` entries.
-  """
-  @spec messages(binary(), paging_options: PagingOptions.t()) :: [Message.t()]
-  def messages(direction, options) when direction == "from-rollup" do
-    do_messages(:from_l2, options)
-  end
-
-  def messages(direction, options) when direction == "to-rollup" do
-    do_messages(:to_l2, options)
-  end
-
-  # Executes the query to fetch cross-chain messages based on the specified direction.
-  #
-  # This function constructs and executes a query to retrieve messages either sent
-  # to or from the rollup layer, applying pagination options. These options dictate
-  # not only the number of items to retrieve but also how many items to skip from
-  # the top.
-  #
-  # ## Parameters
-  # - `direction`: Can be either `:from_l2` or `:to_l2`, indicating the direction of the messages.
-  # - `options`: A keyword list which may contain `paging_options` specifying pagination details
-  #
-  # ## Returns
-  # - A list of `Explorer.Chain.Arbitrum.Message` entries matching the specified direction.
-  @spec do_messages(:from_l2 | :to_l2, paging_options: PagingOptions.t()) :: [Message.t()]
-  defp do_messages(direction, options) do
-    base_query =
-      from(msg in Message,
-        where: msg.direction == ^direction,
-        order_by: [desc: msg.message_id]
-      )
-
-    paging_options = Keyword.get(options, :paging_options, Chain.default_paging_options())
-
-    query =
-      base_query
-      |> page_messages(paging_options)
-      |> limit(^paging_options.page_size)
-
-    select_repo(@api_true).all(query)
-  end
-
-  defp page_messages(query, %PagingOptions{key: nil}), do: query
-
-  defp page_messages(query, %PagingOptions{key: {id}}) do
-    from(msg in query, where: msg.message_id < ^id)
-  end
-
-  @doc """
-    Retrieves a list of relayed L1 to L2 messages that have been completed.
-
-    ## Parameters
-    - `options`: A keyword list which may contain `paging_options` specifying pagination details
-
-    ## Returns
-    - A list of `Explorer.Chain.Arbitrum.Message` representing relayed messages from L1 to L2 that have been completed.
-  """
-  @spec relayed_l1_to_l2_messages(paging_options: PagingOptions.t()) :: [Message.t()]
-  def relayed_l1_to_l2_messages(options) do
-    paging_options = Keyword.get(options, :paging_options, Chain.default_paging_options())
-
-    query =
-      from(msg in Message,
-        where: msg.direction == :to_l2 and not is_nil(msg.completion_transaction_hash),
-        order_by: [desc: msg.message_id],
-        limit: ^paging_options.page_size
-      )
-
-    select_repo(@api_true).all(query)
-  end
 
   @doc """
     Retrieves the total count of rollup batches indexed up to the current moment.
@@ -627,54 +470,6 @@ defmodule Explorer.Chain.Arbitrum.Reader.API do
     Common.get_anytrust_keyset(keyset_hash, api?: true)
   end
 
-  #################################################################################
-  ### Below are functions that implement functionality not specific to Arbitrum ###
-  ### They are candidates for moving to a chain-agnostic module as soon as such ###
-  ### need arises.                                                              ###
-  #################################################################################
-
-  @doc """
-    Retrieves logs from a transaction that match a specific topic.
-
-    Fetches all logs emitted by the specified transaction that have the given topic
-    as their first topic, ordered by log index.
-
-    ## Parameters
-    - `transaction_hash`: The hash of the transaction to fetch logs from
-    - `topic0`: The first topic to filter logs by
-
-    ## Returns
-    - A list of matching logs ordered by index, or empty list if none found
-  """
-  @spec transaction_to_logs_by_topic0(Hash.Full.t(), binary()) :: [Log.t()]
-  def transaction_to_logs_by_topic0(transaction_hash, topic0) do
-    Chain.log_with_transactions_query()
-    |> where([log, transaction], transaction.hash == ^transaction_hash and log.first_topic == ^topic0)
-    |> order_by(asc: :index)
-    |> select_repo(@api_true).all()
-  end
-
-  # Builds a query to fetch batch numbers associated with a DA blob hash.
-  #
-  # This function constructs an Ecto query to retrieve batch numbers from the
-  # arbitrum_batches_to_da_blobs table that match a specific data blob ID (hash).
-  # The batch numbers are sorted in descending order.
-  #
-  # ## Parameters
-  # - `data_key`: The hash of the data blob to find associated batch numbers for.
-  #
-  # ## Returns
-  # - An Ecto query that can be executed to fetch matching batch numbers.
-  @spec build_batch_numbers_by_data_key_query(binary()) :: Ecto.Query.t()
-  defp build_batch_numbers_by_data_key_query(data_key) do
-    from(
-      link in BatchToDaBlob,
-      where: link.data_blob_id == ^data_key,
-      select: link.batch_number,
-      order_by: [desc: link.batch_number]
-    )
-  end
-
   @doc """
     Retrieves all batch numbers associated with a Data Availability (DA) blob hash
     and the corresponding DA blob description.
@@ -715,5 +510,26 @@ defmodule Explorer.Chain.Arbitrum.Reader.API do
           {:error, :not_found} = error -> error
         end
     end
+  end
+
+  # Builds a query to fetch batch numbers associated with a DA blob hash.
+  #
+  # This function constructs an Ecto query to retrieve batch numbers from the
+  # arbitrum_batches_to_da_blobs table that match a specific data blob ID (hash).
+  # The batch numbers are sorted in descending order.
+  #
+  # ## Parameters
+  # - `data_key`: The hash of the data blob to find associated batch numbers for.
+  #
+  # ## Returns
+  # - An Ecto query that can be executed to fetch matching batch numbers.
+  @spec build_batch_numbers_by_data_key_query(binary()) :: Ecto.Query.t()
+  defp build_batch_numbers_by_data_key_query(data_key) do
+    from(
+      link in BatchToDaBlob,
+      where: link.data_blob_id == ^data_key,
+      select: link.batch_number,
+      order_by: [desc: link.batch_number]
+    )
   end
 end
